chore(review): перенес и структурировал в директорию api
