+++ src/App/api/type/directionProduct/list.ts
@@ -1,0 +1,8 @@
+export type TFiltersApi = {
+  title?: string;
+  categoryId?: number | undefined | null;
+  page?: number;
+  limit?: number;
+  price_min?: number | null;
+  price_max?: number | null;
+};
