+++ src/App/stores/FilterStore/FilterStore.ts
@@ -1,9 +1,9 @@
 import { makeAutoObservable } from "mobx";
-import { TFiltersApi } from "@types";
 import { initialFilters } from "@utils/constants";
 import QueryParamsStore from "@stores/RootStore/QueryParamsStore";
 import QueryFilterConnector from "@stores/QueryFilterConnector";
 import { NavigateFunction } from "react-router-dom";
+import { TFiltersApi } from "@api/type/directionProduct/list";
 
 export default class FilterStore {
   private _fieldTitle: string = "";
