+++ src/App/utils/types.ts
@@ -23,12 +23,3 @@ export type OptionEntity<K = string, V = string> = {
   value: K;
   key: V;
 };
-
-export type TFiltersApi = {
-  title?: string;
-  categoryId?: number | undefined | null;
-  page?: number;
-  limit?: number;
-  price_min?: number | null;
-  price_max?: number | null;
-};
